+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/dc/equations/ClosedBranchDcContext.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.loadflow.simple.dc;
+package com.powsybl.loadflow.simple.dc.equations;
 
 import com.powsybl.iidm.network.Branch;
 import com.powsybl.iidm.network.Bus;
