+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/dc/equations/DcEquationSystem.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.loadflow.simple.dc;
+package com.powsybl.loadflow.simple.dc.equations;
 
 import com.powsybl.iidm.network.Branch;
 import com.powsybl.iidm.network.Bus;
@@ -17,10 +17,12 @@ import java.util.List;
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class DcEquationSystemMaker implements EquationSystemMaker {
+public final class DcEquationSystem {
 
-    @Override
-    public EquationSystem make(NetworkContext networkContext, EquationContext equationContext) {
+    private DcEquationSystem() {
+    }
+
+    public static EquationSystem create(NetworkContext networkContext, EquationContext equationContext) {
         List<EquationTerm> equationTerms = new ArrayList<>();
         List<VariableUpdate> variableUpdates = new ArrayList<>();
