--- simple-loadflow/src/main/java/com/powsybl/loadflow/simple/equations/EquationSystemMaker.java
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.loadflow.simple.equations;
-
-import com.powsybl.loadflow.simple.network.NetworkContext;
-
-/**
- * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
- */
-public interface EquationSystemMaker {
-
-    EquationSystem make(NetworkContext networkContext, EquationContext equationContext);
-}
